requirements.txt
amqp==2.5.2
anyjson==0.3.3
apturl==0.5.2
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
billiard==3.6.1.0
blinker==1.4
Brlapi==0.7.0
cachetools==5.2.0
celery==4.2.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.5
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
duplicity==0.8.12.0
entrypoints==0.3
ephem==3.7.7.1
fasteners==0.14.1
filelock==3.7.1
future==0.18.2
google-api-core==2.10.1
google-api-python-client==2.61.0
google-auth==2.11.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==2.8
keyring==18.0.1
kombu==4.6.7
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
netifaces==0.10.4
numpy==1.23.2
oauthlib==3.1.0
olefile==0.46
pandas==1.5.0
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.5.2
protobuf==4.21.6
psycopg2-binary==2.9.3
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.8.0
pycups==1.9.73
PyGObject==3.36.0
pygsheets==2.0.5
PyJWT==1.7.1
pylance==0.0.3
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-memcached==1.58
pytz==2022.2.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
systemd-python==234
toml==0.10.2
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
usb-creator==0.3.7
vine==1.3.0
virtualenv==20.15.1
wadllib==1.3.3
xkit==0.0.0
